move opensea to collectible contract information
